third_party/blink/web_tests/wpt_internal/css/css-text/parsing/word-boundary-detection-auto.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing word-boundary-detection with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#word-boundary-detection">
<meta name="assert" content="word-boundary-detection supports the full grammar 'normal | manual | auto(<lang>) '.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
/*
 * There are valid/invalid/computed tests in `wpt/css/css-text/parsing`, but
 * they can't test the `auto(lang)` values because whether it's valid or not
 * depends on supported languages, which is UA dependent. See the comments in
 * these tests.
 *
 * These tests complement the `auto(lang)` part.
 */
test_valid_value("word-boundary-detection", "auto(ja)");

test_invalid_value("word-boundary-detection", "auto(en)");
test_invalid_value("word-boundary-detection", "auto(zh)");

test_computed_value("word-boundary-detection", "auto(ja)");
</script>
</body>
</html>
